+++ lisa-core/src/main/java/it/unive/lisa/analysis/dataflow/AvailableExpressions.java
@@ -1,9 +1,11 @@
-package it.unive.lisa.analysis.impl.dataflow;
+package it.unive.lisa.analysis.dataflow;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
 
 import it.unive.lisa.analysis.ScopeToken;
 import it.unive.lisa.analysis.SemanticException;
-import it.unive.lisa.analysis.dataflow.DataflowElement;
-import it.unive.lisa.analysis.dataflow.DefiniteForwardDataflowDomain;
 import it.unive.lisa.analysis.representation.DomainRepresentation;
 import it.unive.lisa.analysis.representation.StringRepresentation;
 import it.unive.lisa.program.cfg.ProgramPoint;
@@ -16,9 +18,6 @@ import it.unive.lisa.symbolic.value.Skip;
 import it.unive.lisa.symbolic.value.TernaryExpression;
 import it.unive.lisa.symbolic.value.UnaryExpression;
 import it.unive.lisa.symbolic.value.ValueExpression;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
 
 /**
  * An implementation of the available expressions dataflow analysis, that
