+++ lisa-core/src/main/java/it/unive/lisa/analysis/dataflow/ReachingDefinitions.java
@@ -1,9 +1,11 @@
-package it.unive.lisa.analysis.impl.dataflow;
+package it.unive.lisa.analysis.dataflow;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
 
 import it.unive.lisa.analysis.ScopeToken;
 import it.unive.lisa.analysis.SemanticException;
-import it.unive.lisa.analysis.dataflow.DataflowElement;
-import it.unive.lisa.analysis.dataflow.PossibleForwardDataflowDomain;
 import it.unive.lisa.analysis.representation.DomainRepresentation;
 import it.unive.lisa.analysis.representation.PairRepresentation;
 import it.unive.lisa.analysis.representation.StringRepresentation;
@@ -12,9 +14,6 @@ import it.unive.lisa.symbolic.SymbolicExpression;
 import it.unive.lisa.symbolic.value.Identifier;
 import it.unive.lisa.symbolic.value.OutOfScopeIdentifier;
 import it.unive.lisa.symbolic.value.ValueExpression;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
 
 /**
  * An implementation of the reaching definition dataflow analysis.
