+++ lisa-core/src/main/java/it/unive/lisa/analysis/heap/MonolithicHeap.java
@@ -1,8 +1,10 @@
-package it.unive.lisa.analysis.impl.heap;
+package it.unive.lisa.analysis.heap;
+
+import java.util.Collections;
+import java.util.List;
 
 import it.unive.lisa.analysis.Lattice;
 import it.unive.lisa.analysis.SemanticException;
-import it.unive.lisa.analysis.heap.BaseHeapDomain;
 import it.unive.lisa.analysis.lattices.ExpressionSet;
 import it.unive.lisa.analysis.representation.DomainRepresentation;
 import it.unive.lisa.analysis.representation.StringRepresentation;
@@ -17,8 +19,6 @@ import it.unive.lisa.symbolic.value.HeapLocation;
 import it.unive.lisa.symbolic.value.Identifier;
 import it.unive.lisa.symbolic.value.MemoryPointer;
 import it.unive.lisa.symbolic.value.ValueExpression;
-import java.util.Collections;
-import java.util.List;
 
 /**
  * A monolithic heap implementation that abstracts all heap locations to a
