+++ lisa-core/src/main/java/it/unive/lisa/analysis/heap/pointbased/AllocationSites.java
@@ -1,4 +1,4 @@
-package it.unive.lisa.analysis.impl.heap.pointbased;
+package it.unive.lisa.analysis.heap.pointbased;
 
 import it.unive.lisa.analysis.SemanticDomain.Satisfiability;
 import it.unive.lisa.analysis.SemanticException;
