+++ lisa-core/src/main/java/it/unive/lisa/interprocedural/SingleScopeToken.java
@@ -1,4 +1,4 @@
-package it.unive.lisa.interprocedural.impl;
+package it.unive.lisa.interprocedural;
 
 import it.unive.lisa.analysis.ScopeToken;
 import java.util.Objects;
